# protocol/openapi.yaml
openapi: 3.1.0
info:
  title: Rivune API
  version: 1.5.0
  description: >
    Protocol version 20 between Rivune clients and a self-hosted Rivune Server.
    While global maintenance mode is enabled, authenticated non-admin application
    operations return the reusable 503 MaintenanceMode response. Authentication,
    session recovery, health, discovery, and administrator operations remain available.
  license:
    name: Apache License 2.0
    identifier: Apache-2.0
servers:
  - url: /api/v1
paths:
...
  /playback/prepare:
    post:
      summary: Prepare one opaque playback source
      operationId: preparePlayback
      description: Resolves and caches the media decision for one selected source reference without creating a playback session. Concurrent inspection of the same media is deduplicated.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaybackPrepareInput"
      responses:
        "200":
          description: Cached playback decision for the selected source.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackPreparation"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "410":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/PlaybackUnprocessable"
        "502":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/PlaybackCapacity"
        "507":
          $ref: "#/components/responses/Error"
  /playback/resolve:
    post:
      summary: Create a playback session from an opaque source reference
      operationId: resolvePlayback
      description: Reuses the cached preparation for a session-bound source reference and creates a short-lived playback session. It does not refetch stream resources. Web media URLs and private headers remain server-side; mode=external returns the authorized provider handoff URL for the selected external player.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaybackResolveInput"
      responses:
        "201":
          description: A short-lived playback session for the selected source.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackSession"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "410":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/PlaybackUnprocessable"
        "502":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/PlaybackCapacity"
        "507":
          $ref: "#/components/responses/Error"
  /playback/sessions/{sessionId}:
    parameters:
      - name: sessionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      summary: Stop the caller's playback session
      operationId: stopPlayback
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Playback session and media processing stopped.
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    setupToken:
      type: http
      scheme: bearer
      description: One-time bootstrap secret configured by the server administrator.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >
        Short-lived opaque Rivune access token. Every session has immutable global_admin or category
        authorization scope. Only administrator password login may issue global_admin. Category identity
        is derived and revalidated from server-owned session and device state; client category fields are
        resource identifiers to authorize, never claims. Category moves revoke incompatible sessions. After profile
        selection, web clients must also present that selection's opaque X-Rivune-Profile-Context capability on
        authenticated requests other than authentication, profile listing, profile selection, and profile avatar reads.
    demoCookie:
      type: apiKey
      in: cookie
      name: rivune_demo
      description: Opaque pre-setup demonstration cookie issued with HttpOnly, SameSite=Strict, and Path=/api/v1; Secure for direct or forwarded HTTPS requests.
    playbackToken:
      type: apiKey
      in: query
      name: token
      description: Opaque short-lived token returned only inside playback asset URLs.
    calendarToken:
      type: apiKey
      in: query
      name: token
      description: >
        Long-lived opaque bearer secret embedded in a private calendar URL. Its rivune_cal_ prefixed value contains
        32 bytes from a cryptographically secure random generator; the server persists only its SHA-256 digest and
        never logs or includes the token in an error. The complete secret is disclosed only when the subscription is
        created or rotated, is invalidated immediately by rotation or revocation, and must be the request's only
        query parameter with exactly one value. This credential is independent of bearerAuth.
...
  responses:
    Error:
      description: Structured API error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    PlaybackUnprocessable:
      description: Playback input is invalid, transcoding is disabled, or a required client processing capability was not announced.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PlaybackUnprocessableError"
    PlaybackCapacity:
      description: Every server media-processing slot is occupied, or global maintenance mode blocks the playback request.
      headers:
        Retry-After:
          required: true
          description: Suggested delay in seconds before retrying playback preparation or resolution.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            anyOf:
              - $ref: "#/components/schemas/PlaybackCapacityError"
              - $ref: "#/components/schemas/MaintenanceModeError"
...
  schemas:
...
    MaintenanceModeError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              const: maintenance_mode
            message:
              type: string
              maxLength: 500
            publicMessage:
              description: Optional administrator-provided public message; clients localize their fallback when absent.
              type: string
              maxLength: 500
...
    PlaybackResolveInput:
      type: object
      additionalProperties: false
      required: [sourceRef]
      properties:
        sourceRef:
          type: string
          minLength: 16
          maxLength: 128
        titleId:
          type: string
          maxLength: 128
        preferredAudioTrack:
          type: integer
          minimum: 0
          maximum: 10000
        preferredSubtitleId:
          type: string
          maxLength: 128
        startSeconds:
          description: Whole-second playback offset applied to processed media.
          type: integer
          minimum: 0
          maximum: 604800
...
    PlaybackPrepareInput:
      type: object
      additionalProperties: false
      required: [sourceRef]
      properties:
        sourceRef:
          type: string
          minLength: 16
          maxLength: 128
        startSeconds:
          description: Whole-second playback offset applied to processed media.
          type: integer
          minimum: 0
          maximum: 604800
    PlaybackPreparation:
      type: object
      additionalProperties: false
      required: [sourceRef, mode, protocol, subtitleCount, expiresAt]
      properties:
        sourceRef:
          type: string
          minLength: 16
          maxLength: 128
        mode:
          type: string
          enum: [direct, remux, transcode_audio, transcode, youtube, external]
        protocol:
          type: string
        decision:
          $ref: "#/components/schemas/PlaybackDecision"
        container:
          type: string
        media:
          $ref: "#/components/schemas/PlaybackMediaInspection"
        subtitleCount:
          type: integer
          minimum: 0
        expiresAt:
          type: string
          format: date-time
    PlaybackDecision:
      type: object
      additionalProperties: false
      required: [reason, videoAction, audioAction, subtitleAction, toneMapping]
      properties:
        reason:
          type: string
          enum: [direct_supported, remux_required, audio_transcode_required, video_transcode_required, subtitle_burn_required]
        videoAction:
          type: string
          enum: [copy, transcode]
        audioAction:
          type: string
          enum: [copy, transcode]
        subtitleAction:
          type: string
          enum: [none, external, copy, burn]
        toneMapping:
          type: boolean
        source:
          $ref: "#/components/schemas/PlaybackDecisionSource"
        target:
          $ref: "#/components/schemas/PlaybackDecisionTarget"
    PlaybackDecisionSource:
      type: object
      additionalProperties: false
      properties:
        container:
          type: string
        videoCodec:
          type: string
        audioCodec:
          type: string
        height:
          type: integer
          minimum: 1
        videoBitrateKbps:
          type: integer
          minimum: 1
        hdrFormat:
          type: string
    PlaybackDecisionTarget:
      type: object
      additionalProperties: false
      properties:
        protocol:
          type: string
        container:
          type: string
        videoCodec:
          type: string
        audioCodec:
          type: string
        height:
          type: integer
          minimum: 1
          maximum: 4320
        videoBitrateKbps:
          type: integer
          minimum: 64
          maximum: 200000
...
    PlaybackSession:
      type: object
      additionalProperties: false
      required: [id, selectedSourceId, sources, subtitles, providerErrors, expiresAt]
      properties:
        id:
          type: string
          format: uuid
        selectedAudioTrack:
          type: integer
          minimum: 0
          maximum: 10000
        selectedSubtitleId:
          type: string
          maxLength: 128
        selectedSourceId:
          type: string
        sources:
          type: array
          items:
            $ref: "#/components/schemas/PlaybackSource"
        subtitles:
          type: array
          items:
            $ref: "#/components/schemas/PlaybackSubtitle"
        providerErrors:
          type: array
          items:
            $ref: "#/components/schemas/PlaybackProviderError"
        expiresAt:
          type: string
          format: date-time
...
    PlaybackSource:
      type: object
      additionalProperties: false
      required: [id, addonId, manifestId, mode, protocol, compatible]
      properties:
        id:
          type: string
        addonId:
          type: string
          format: uuid
        manifestId:
          type: string
        name:
          type: string
        title:
          type: string
        mode:
          type: string
          enum: [direct, remux, transcode_audio, transcode, youtube, external]
        decision:
          $ref: "#/components/schemas/PlaybackDecision"
        url:
          type: string
          description: Same-server playback asset URL containing a short-lived opaque token, except mode=external where this is the authorized provider handoff URL.
        ytId:
          type: string
        infoHash:
          type: string
        fileIndex:
          type: integer
          minimum: 0
        protocol:
          type: string
        container:
          type: string
        compatible:
          type: boolean
        media:
          $ref: "#/components/schemas/PlaybackMediaInspection"
    PlaybackMediaInspection:
      type: object
      additionalProperties: false
      required: [videoTracks, audioTracks, subtitleTracks]
      properties:
        container:
          type: string
        durationSeconds:
          type: number
          minimum: 0
        hdrFormat:
          type: string
        videoTracks:
          type: array
          items:
            $ref: "#/components/schemas/PlaybackMediaTrack"
        audioTracks:
          type: array
          items:
            $ref: "#/components/schemas/PlaybackMediaTrack"
        subtitleTracks:
          type: array
          items:
            $ref: "#/components/schemas/PlaybackMediaTrack"
    PlaybackMediaTrack:
      type: object
      additionalProperties: false
      required: [index, type, codec]
      properties:
        index:
          type: integer
          minimum: 0
        type:
          type: string
          enum: [video, audio, subtitle]
        codec:
          type: string
        profile:
          type: string
        language:
          type: string
        title:
          type: string
        forced:
          type: boolean
        width:
          type: integer
          minimum: 1
        height:
          type: integer
          minimum: 1
        channels:
          type: integer
          minimum: 1
    PlaybackSubtitle:
      type: object
      additionalProperties: false
      required: [id, addonId, manifestId]
      properties:
        id:
          type: string
        addonId:
          type: string
          format: uuid
        manifestId:
          type: string
        language:
          type: string
        forced:
          type: boolean
        default:
          type: boolean
        delivery:
          type: string
          enum: [external, burn]
        url:
          type: string
          description: Same-server playback asset URL containing the playback token. Omitted when delivery=burn because the selected subtitle is embedded in the video output.
    PlaybackProviderError:
      type: object
      additionalProperties: false
      required: [addonId, manifestId, code, message]
      properties:
        addonId:
          type: string
          format: uuid
        manifestId:
          type: string
        code:
          type: string
        message:
          type: string
...
    PlaybackUnprocessableError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              enum: [invalid_playback_request, playback_source_unsupported, playback_transcoding_disabled, playback_client_capability_missing]
            message:
              type: string
    PlaybackCapacityError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              const: playback_capacity_reached
            message:
              type: string
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...